Restructured/re-abstracted Demo
