+++ demo/common/.gitignore
@@ -1,0 +1,23 @@
+/data
+/data*
+/demo
+/demo*
+/config*
+*.tgz 
+src/main/resources/mesh-demo.zip
+/target/
+mesh-dump.zip
+data-nodeA
+orient-plugins/studio*.zip
+/elasticsearch
+mesh-ui-config.js
+data_*
+restore*.sh
+migrated.sh
+keystore.jceks
+mesh.yml
+data-nodeB
+config-nodeA
+config-nodeB
+orientdb-plugins
+config
