+++ demo/common/src/main/java/com/gentics/mesh/demo/DemoZipHelper.java
