+++ demo/common/src/main/java/com/gentics/mesh/demo/verticle/DemoVerticle.java
