+++ demo/common/src/main/resources/data/groups.json
@@ -1,0 +1,6 @@
+{
+  "data": [
+    { "uuid": "df81c23d9ff1450081c23d9ff195005e", "name": "Client Group", "roles": ["Client Role", "admin"] },
+    { "uuid": "e728c2fc5a0146a2a8c2fc5a0186a28d", "name": "Editor Group", "roles": ["Editor Role", "admin"] }
+  ]
+}
