+++ demo/common/src/main/resources/data/images/sources.txt
@@ -1,0 +1,11 @@
+embraer-legacy-600.jpg                  - https://en.wikipedia.org/wiki/File:Embraer_legacy_600_g-irsh_arp.jpg                                                                                        - public domain
+ford-gt.jpg                             - https://commons.wikimedia.org/wiki/File:2005_Ford_GT_pic1.JPG                                                                                               - public domain
+gulfstream-g550.jpg                     - https://commons.wikimedia.org/wiki/File:Gulfstream.g550.cs-dkf.climb.arp.jpg                                                                                - public domain
+koenigsegg-ccx.jpg                      - https://commons.wikimedia.org/wiki/File:Koenigsegg_CCX_parked.jpg                                                                                           - public domain
+sts.jpg                                 - https://commons.wikimedia.org/wiki/File:STS-133_Last_towback_of_Space_Shuttle_Discovery.jpg                                                                 - public domain
+tesla-roadster.jpg                      - https://commons.wikimedia.org/wiki/File:Tesla_Roadster_--_02-11-2011.jpg                                                                                    - public domain
+trabant.jpg                             - http://www.publicdomainpictures.net/view-image.php?image=117790&picture=trabant-601s-auto                                                                   - public domain
+Yacht_Indian_Empress.jpg                - https://commons.wikimedia.org/wiki/File:Yacht_Indian_Empress.jpg                                                                                            - public domain
+yacht-pelorus.jpg                       - https://commons.wikimedia.org/wiki/File:Pelorus_at_Nassau.JPG                                                                                               - public domain
+maize-blue-solar-car.jpg                - https://commons.wikimedia.org/wiki/File:Maize_and_Blue_solar_car,_University_of_Michigan,_1993_-_Museum_of_Science_and_Industry_(Chicago)_-_DSC06862.JPG    - public domain
+delorean-dmc-12.jpg                     - http://www.publicdomainpictures.net/view-image.php?image=183780&picture=1981-de-lorean-dmc-12                                                               - public domain
