+++ demo/common/src/main/resources/data/projects.json
@@ -1,0 +1,5 @@
+{
+  "data": [
+    { "uuid": "217f8c981ada4642bf8c981adaa642c3", "name": "demo" }
+  ]
+}
