+++ demo/common/src/main/resources/data/roles.json
@@ -1,0 +1,6 @@
+{
+  "data": [
+    { "uuid": "ddcddceb33d648318ddceb33d618314f", "name": "Client Role" },
+    { "uuid": "6754d83b086244d894d83b086204d87f", "name": "Editor Role" }
+  ]
+}
