+++ demo/common/src/main/resources/data/schemas/category.json
@@ -1,0 +1,22 @@
+{
+  "name" : "category",
+  "container" : true,
+  "displayField": "name",
+  "segmentField": "slug",
+  "fields" : [ {
+    "name" : "name",
+    "label" : "Name",
+    "required" : true,
+    "type" : "string"
+  }, {
+    "name" : "slug",
+    "label" : "Slug",
+    "required" : true,
+    "type" : "string"
+  }, {
+    "name" : "description",
+    "label" : "Description",
+    "required" : false,
+    "type" : "string"
+  } ]
+}
