+++ demo/common/src/main/resources/data/schemas/vehicle.json
@@ -1,0 +1,56 @@
+{
+  "name" : "vehicle",
+  "container" : false,
+  "displayField": "name",
+  "segmentField": "slug",
+  "urlFields": [ "niceUrl" ],
+  "fields" : [ {
+    "name" : "slug",
+    "label" : "Slug",
+    "required" : true,
+    "type" : "string"
+  }, {
+    "name" : "niceUrl",
+    "label" : "Nice URL",
+    "required" : false,
+    "type" : "string"
+  }, {
+    "name" : "name",
+    "label" : "Name",
+    "required" : true,
+    "type" : "string"
+  }, {
+    "name" : "weight",
+    "label" : "Weight",
+    "required" : false,
+    "type" : "number"
+  }, {
+    "name" : "SKU",
+    "label" : "Stock Keeping Unit",
+    "required" : false,
+    "type" : "number"
+  }, {
+    "name" : "price",
+    "label" : "Price",
+    "required" : false,
+    "type" : "number"
+  }, {
+    "name" : "stocklevel",
+    "label" : "Stock Level",
+    "required" : false,
+    "type" : "number"
+  }, {
+    "name" : "description",
+    "label" : "Description",
+    "required" : false,
+    "type" : "html"
+  }, {
+      "name": "vehicleImage",
+      "label": "Vehicle Image",
+      "type": "node",
+      "allow": [
+        "vehicleImage"
+      ]
+    }
+   ]
+}
