+++ demo/common/src/main/resources/data/schemas/vehicleImage.json
@@ -1,0 +1,22 @@
+{
+  "name" : "vehicleImage",
+  "container" : false,
+  "displayField": "name",
+  "segmentField": "image",
+  "fields" : [ {
+    "name" : "name",
+    "label" : "Name",
+    "required" : true,
+    "type" : "string"
+  }, {
+    "name" : "altText",
+    "label" : "Alternative Text",
+    "required" : false,
+    "type" : "string"
+  }, {
+    "name" : "image",
+    "label" : "Image",
+    "required" : false,
+    "type" : "binary"
+  } ]
+}
