+++ demo/default/pom.xml
@@ -9,7 +9,7 @@
 
 	<parent>
 		<groupId>com.gentics.mesh</groupId>
-		<artifactId>mesh-demo-common</artifactId>
+		<artifactId>mesh-demos</artifactId>
 		<version>1.8.6-SNAPSHOT</version>
 	</parent>
 
@@ -32,6 +32,11 @@
 	</dependencyManagement>
 
 	<dependencies>
+		<dependency>
+			<groupId>com.gentics.mesh</groupId>
+			<artifactId>mesh-demo-common</artifactId>
+			<version>${project.version}</version>
+		</dependency>
 		<dependency>
 			<groupId>com.gentics.mesh</groupId>
 			<artifactId>mesh-server</artifactId>
