+++ demo/pom.xml
@@ -2,10 +2,10 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<artifactId>mesh-demo-common</artifactId>
+	<artifactId>mesh-demos</artifactId>
 
 	<packaging>pom</packaging>
-	<name>Mesh - Common Demo functions</name>
+	<name>Mesh - Demo modules</name>
 
 	<parent>
 		<groupId>com.gentics.mesh</groupId>
@@ -17,6 +17,7 @@
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 	</properties>
 	<modules>
+		<module>common</module>
 		<module>default</module>
 	</modules>
