+++ tests/tests-demo/.gitignore
@@ -1,0 +1,1 @@
+/hmac_store.jceks
