+++ tests/tests-demo/README.md
@@ -1,0 +1,3 @@
+# Gentics Mesh - Core Tests
+
+Attention: Plugin tests may require pre-build test plugins in order to be executed locally. The build for those plugins can be invoked using the `build-test-plugins.sh` script.
